move FuncTranslator internals into func submodule
