+++ crates/wasmi/src/engine/translator/func/instr_encoder.rs
@@ -1,21 +1,25 @@
-use super::{
-    comparator::TryIntoCmpSelectInstr,
-    relink_result::RelinkResult as _,
-    utils::{FromProviders as _, WasmInteger},
-    visit_register::VisitInputRegisters as _,
-    BumpFuelConsumption as _,
-    FuelInfo,
-    LabelRef,
-    LabelRegistry,
-    LogicalizeCmpInstr,
-    NegateCmpInstr,
-    TryIntoCmpBranchFallbackInstr,
-    TryIntoCmpBranchInstr,
-    TypedProvider,
-};
 use crate::{
     core::{FuelCostsProvider, UntypedVal, ValType},
-    engine::translator::{stack::RegisterSpace, ValueStack},
+    engine::translator::{
+        func::{
+            stack::RegisterSpace,
+            utils::FromProviders as _,
+            FuelInfo,
+            LabelRef,
+            LabelRegistry,
+            LogicalizeCmpInstr,
+            NegateCmpInstr,
+            TryIntoCmpBranchFallbackInstr,
+            TryIntoCmpBranchInstr,
+            TryIntoCmpSelectInstr,
+            TypedProvider,
+            ValueStack,
+        },
+        relink_result::RelinkResult as _,
+        utils::WasmInteger,
+        visit_register::VisitInputRegisters as _,
+        BumpFuelConsumption as _,
+    },
     ir::{
         BoundedRegSpan,
         BranchOffset,
