+++ crates/wasmi/src/engine/translator/func/provider.rs
