+++ crates/wasmi/src/engine/translator/func/stack/provider.rs
@@ -1,5 +1,5 @@
 use super::{LocalRefs, RegisterAlloc, TypedVal};
-use crate::{engine::translator::PreservedLocal, ir::Reg, Error};
+use crate::{engine::translator::func::PreservedLocal, ir::Reg, Error};
 use alloc::vec::Vec;
 use arrayvec::ArrayVec;
